[downloadfile-unvendor] Move system package to exec package. Move FileExists to generic utils
