+++ pkg/exec/exec.go
@@ -1,4 +1,4 @@
-package system
+package exec
 
 import (
 	"os"
@@ -37,14 +37,4 @@ func RunCommandPipe(command string, args []string) error {
 	proc.Stderr = os.Stderr
 
 	return proc.Run()
-}
-
-// FileExists checks a file's existence
-func FileExists(name string) bool {
-	if _, err := os.Stat(name); err != nil {
-		if os.IsNotExist(err) {
-			return false
-		}
-	}
-	return true
 }
