+++ pkg/util/files.go
@@ -291,3 +291,13 @@ func CopyDir(src string, dst string) error {
 
 	return nil
 }
+
+// FileExists checks a file's existence
+func FileExists(name string) bool {
+	if _, err := os.Stat(name); err != nil {
+		if os.IsNotExist(err) {
+			return false
+		}
+	}
+	return true
+}
